#3 - 리팩토링 : 프로젝트 구조 변경
해당 프로그램의 경우 주요 클래스를 정의하고, 해당 클래스를 상속해가면서 기능을 구현한다.

주요 기능 부분과 상속된 부분을 구분하고, import 과정을 간략화하기 위해 프로젝트 구조를 변경한다.
