+++ Aria/evaluations.py
@@ -1,8 +1,8 @@
 from Aria.core.Variable import Variable
-from Aria.core.Utils import as_varialbe, as_array
+from Aria.core.Utils import as_variable, as_array
 
 def accuracy(y, t):
-  y, t = as_varialbe(y), as_varialbe(t)
+  y, t = as_variable(y), as_variable(t)
 
   pred = y.data.argmax(axis=1).reshape(t.shape)
   result = (pred == t.data)
